// src/types/type.ts
export type ActiveElement = {
  name: string;
  value: string;
  icon: string;
} | null;